+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Return `Response` from `send_object` instead of `Option<Response>`.
+
 ## [0.11.0] - 2025-03-19
 
 ### Added